refactor: fixing eslint errors in multiple files
Many files were left with small issues that eslint was showing
as errors during the npm build and in the browser console.
They didn't prevent the package from being build but they
were annoying.

Where I could not find a way to fix the error I turned off
the eslint rule for a specific file.