refactor: replace DateTimePicker with MaskedIinput
The DateTimePicker from react-widgets looks really nice but
I couldn't make it play nice with accessibility so it just has
been replaced with MaskedInput.

Additionaly, the change required refactoring the following:
- validation: dateAndTime and laterThan were added
- single Event component: using moment to correctly parse
and format the start and end dates

Also, in this commit a removing <div> tag around <Event>
in Events page helped to remove the warning: Each child
in an array or iterator should have a unique "key" prop.